+++ src/component/CardItem.jsx
@@ -3,9 +3,9 @@ Modal,Box,Typography, NativeSelect,InputLabel
 ,FormControl} from "@mui/material";
 import { Icon } from '@iconify/react';
 import { PizzaArray } from "./PizzaArray";
-import { CrustArray,SizeArray } from "./AddOnArray";
+import { CrustArray,SizeArray } from "../itemArrays/AddOnArray";
 import React from "react";
-import "./layout.css"
+import "../style/layout.css"
 
 function CardItem()
 {
